relocate kqueue
