+++ CMakeLists.txt
@@ -4,15 +4,28 @@ project(echo_test)
 
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11 -g")
 
-add_library(common STATIC
-    address.cpp
-    cap_read.cpp
-    cap_write.cpp
-    epoll.cpp
-    pipe.cpp
-    socket.cpp
-    throw_error.cpp
-)
+IF(UNIX)
+    IF(APPLE)
+    add_library(common STATIC
+        address.cpp
+        cap_read.cpp
+        cap_write.cpp
+        kqueue.cpp
+        socket.cpp
+        throw_error.cpp
+        )
+    ELSE(APPLE)
+    add_library(common STATIC
+        address.cpp
+        cap_read.cpp
+        cap_write.cpp
+        epoll.cpp
+        socket.cpp
+        throw_error.cpp
+        )
+    ENDIF(APPLE)
+ENDIF(UNIX)
+
 
 add_executable(echo_server
     main_echo_server.cpp
